requirements.txt
setuptools>=39.1.0
python-dateutil~=2.8.1
paramiko~=2.7.2
pandas~=1.1.2